requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# UPE specific dependencies
python-pptx==0.6.23
python-docx==1.1.2
openpyxl==3.1.5
reportlab==4.2.2

# Additional dependencies
pydantic>=2.6.0
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
markdown==3.5.1
